refactor: fixe static analysis errors
Remove unused classe(s) and classe method(s) and dead code.

Rename class methods to use camelCase.

Merge constants into bootstrap script.

Use builtin number_format() function instead of using CUtils::format_Number().

Fix code style to follow PSR-12 rules.

chore: bump phpstan version and config
  Bump phpstan version and update config.

chore: add error types exclusions
  Add exclusions for some error types in PHPStan configuration.

chore: add phpstan doctrine extension
  Add PHPStan Doctrine extension.

+++ core/Db/DBPagination.php
@@ -158,7 +158,7 @@ class DBPagination
         $paginated = new Paginator($queryBuilder, fetchJoinCollection: true);
 
         $this->filteredRow = count($paginated);
-        $this->paginationMax = ceil($this->filteredRow / $this->limit);
+        $this->paginationMax = (int) ceil($this->filteredRow / $this->limit);
 
         return $paginated;
     }
